requirements.txt
# Core Frameworks
Django==4.2.15
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-debug-toolbar==5.1.0
django-silk==5.3.2
django-grappelli==4.0.1

# Database & Caching
psycopg2-binary==2.9.10
redis==5.2.1
django-redis==5.4.0
django-celery-beat==2.8.0
django_celery_results==2.6.0
celery==5.5.1
kombu==5.5.3

# Machine Learning & Data Processing
numpy==1.26.3
pandas==2.2.1
scikit-learn==1.4.1
tensorflow==2.16.2
tensorboard==2.16.0
tensorflow-hub==0.16.1
joblib==1.4.2
h5py==3.13.0

# API & Web Frameworks
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
aiohttp==3.11.16
aiosqlite==0.19.0
uvloop==0.19.0

# Testing & Development Tools
pytest==7.4.4
pytest-asyncio==0.25.2
pytest-django==4.7.1
pytest-cov==4.1.0
coverage==7.4.0
black==23.12.1
flake8==6.1.0
isort==5.13.2
mypy==1.8.0
bandit==1.7.6
safety==2.3.4
openai==0.27.8
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver_manager==4.0.1
pdfkit==1.0.0
reportlab==4.0.7
stripe==7.10.0
numpy==1.24.3
scikit-learn==1.3.0
tweepy==4.13.0
pandas==2.0.3
redis==5.0.1
celery==5.3.4
python-dateutil==2.8.2
pytz==2023.3
stripe==7.10.0
qrcode==7.4.2
Pillow==10.0.0
scikit-learn==1.3.04

# Monitoring & Logging
sentry-sdk==1.47.0
prometheus-client==0.20.0

# Natural Language Processing
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
nltk==3.9.1
transformers==4.51.3
huggingface-hub==0.30.2
sentencepiece==0.2.0

# Authentication & Security
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
oauthlib==3.2.2
requests-oauthlib==2.0.0

# Email & Messaging
aioimaplib==2.0.1
IMAPClient==3.0.1
email_validator==2.2.0

# File Processing & Storage
pillow==10.2.0
pdfminer.six==20250327
pdfplumber==0.11.6
python-docx==1.1.2
weasyprint==60.0
jinja2==3.1.4

# Utilities & Tools
python-dotenv==1.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0


# Machine Learning dependencies
joblib==1.4.2
scikit-learn==1.4.1
numpy==1.26.3
pandas==2.2.1
